+++ emrichen/input/__init__.py
@@ -1,0 +1,14 @@
+from .json import load_json
+from .yaml import load_yaml
+
+PARSERS = {
+    'yaml': load_yaml,
+    'json': load_json,
+}
+
+
+def parse(data, format):
+    if format in PARSERS:
+        return PARSERS[format](data)
+    else:
+        raise ValueError(f'No parser for format {format}')
