+++ emrichen/input/yaml.py
@@ -1,10 +1,9 @@
-import json
 from collections import OrderedDict
 from functools import partial
 
 import yaml
 
-from .tags.base import tag_registry
+from emrichen.tags.base import tag_registry
 
 
 def _construct_tagless_yaml(loader, node):
@@ -40,29 +39,5 @@ class RichLoader(yaml.SafeLoader):
         return OrderedDict(loader.construct_pairs(node))
 
 
-def _hydrate_json_object(pairs):
-    if len(pairs) == 1:
-        key, data = pairs[0]
-        if key.startswith('!') and key[1:] in tag_registry:
-            tag = tag_registry[key[1:]]
-            return tag(data)
-    return OrderedDict(pairs)
-
-
-def json_load_or_loads(str_or_stream, **kwargs):
-    if hasattr(str_or_stream, 'read'):
-        str_or_stream = str_or_stream.read()
-    return json.loads(str_or_stream, **kwargs)
-
-
-PARSERS = {
-    'yaml': lambda data: list(yaml.load_all(data, Loader=RichLoader)),
-    'json': lambda data: [json_load_or_loads(data, object_pairs_hook=_hydrate_json_object)],
-}
-
-
-def parse(data, format):
-    if format in PARSERS:
-        return PARSERS[format](data)
-    else:
-        raise ValueError(f'No parser for format {format}')
+def load_yaml(data):
+    return list(yaml.load_all(data, Loader=RichLoader))
